requirements.txt
streamlit
supabase
google-generativeai
google-api-python-client
gTTS
SpeechRecognition
Pillow
python-dotenv